pyproject.toml
[project]
name = "migrations"
version = "0.1.0"
description = "Migrating data from Google Sheets to MongoDB"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "ruff>=0.7.1",
    "uv>=0.4.28",
]
